+++ mypy.ini
@@ -82,7 +82,7 @@ files = fixtures/mypy-stubs,
         src/sentry/integrations/slack/,
         src/sentry/integrations/message_builder.py,
         src/sentry/integrations/utils/atlassian_connect.py,
-        src/sentry/integrations/utils/cleanup.py,
+        src/sentry/integrations/utils/scope.py,
         src/sentry/integrations/utils/code_mapping.py,
         src/sentry/integrations/utils/codecov.py,
         src/sentry/integrations/vsts/,
