+++ src/sentry/integrations/utils/scope.py
